Refactor simulation code
